Repro for an issue when counting size of structs
It assumes 'packed' structs and assumes 12B size although
it likely should be aligned per the widest POD element to
enable natural alignment for the elements. It's not clearly
spec'd in SPIR-V.

Crashes with the CPU driver due to it checking the arg size.
The GPU driver executes it OK (it likely just doesn't check
the size).